refactor: command 패키지에서 converter대신 별도 mapper를 사용하도록 변경
