+++ src/main/java/umc/cockple/demo/domain/exercise/converter/ExerciseConverter.java
@@ -29,113 +29,7 @@ public class ExerciseConverter {
 
     private final FileService fileService;
 
-    // ========== Command 변환 메서드들 ==========
-    public ExerciseCreateDTO.Command toCreateCommand(ExerciseCreateDTO.Request request) {
-        return ExerciseCreateDTO.Command.builder()
-                .date(request.toParsedDate())
-                .startTime(request.toParsedStartTime())
-                .endTime(request.toParsedEndTime())
-                .maxCapacity(request.maxCapacity())
-                .partyGuestAccept(request.allowMemberGuestsInvitation())
-                .outsideGuestAccept(request.allowExternalGuests())
-                .notice(request.notice())
-                .build();
-    }
-
-    public ExerciseCreateDTO.AddrCommand toAddrCreateCommand(ExerciseCreateDTO.Request request) {
-        return ExerciseCreateDTO.AddrCommand.builder()
-                .roadAddress(request.roadAddress())
-                .buildingName(request.buildingName())
-                .latitude(request.latitude())
-                .longitude(request.longitude())
-                .build();
-    }
-
-    public ExerciseGuestInviteDTO.Command toGuestInviteCommand(ExerciseGuestInviteDTO.Request request, Long inviterId) {
-        return ExerciseGuestInviteDTO.Command.builder()
-                .guestName(request.guestName())
-                .gender(request.toParsedGender())
-                .level(request.toParsedLevel())
-                .inviterId(inviterId)
-                .build();
-    }
-
-    public ExerciseUpdateDTO.Command toUpdateCommand(ExerciseUpdateDTO.Request request) {
-        return ExerciseUpdateDTO.Command.builder()
-                .date(request.toParsedDate())
-                .startTime(request.toParsedStartTime())
-                .endTime(request.toParsedEndTime())
-                .maxCapacity(request.maxCapacity())
-                .partyGuestAccept(request.allowMemberGuestsInvitation())
-                .outsideGuestAccept(request.allowExternalGuests())
-                .notice(request.notice())
-                .build();
-    }
-
-    public ExerciseUpdateDTO.AddrCommand toAddrUpdateCommand(ExerciseUpdateDTO.Request request) {
-        if (request.roadAddress() != null || request.buildingName() != null ||
-                request.latitude() != null || request.longitude() != null) {
-            return ExerciseUpdateDTO.AddrCommand.builder()
-                    .roadAddress(request.roadAddress())
-                    .buildingName(request.buildingName())
-                    .latitude(request.latitude())
-                    .longitude(request.longitude())
-                    .build();
-        }
-        return null;
-    }
-
     // ========== Response 변환 메서드들 ==========
-    public ExerciseCreateDTO.Response toCreateResponse(Exercise exercise) {
-        return ExerciseCreateDTO.Response.builder()
-                .exerciseId(exercise.getId())
-                .createdAt(exercise.getCreatedAt())
-                .build();
-    }
-
-    public ExerciseJoinDTO.Response toJoinResponse(MemberExercise memberExercise, Exercise exercise) {
-        return ExerciseJoinDTO.Response.builder()
-                .participantId(memberExercise.getId())
-                .joinedAt(memberExercise.getCreatedAt())
-                .currentParticipants(exercise.getNowCapacity())
-                .build();
-    }
-
-    public ExerciseGuestInviteDTO.Response toGuestInviteResponse(Guest guest, Exercise exercise) {
-        return ExerciseGuestInviteDTO.Response.builder()
-                .guestId(guest.getId())
-                .invitedAt(guest.getCreatedAt())
-                .currentParticipants(exercise.getNowCapacity())
-                .build();
-    }
-
-    public ExerciseCancelDTO.Response toCancelResponse(Exercise exercise, Member member) {
-        return ExerciseCancelDTO.Response.builder()
-                .memberName(member.getMemberName())
-                .currentParticipants(exercise.getNowCapacity())
-                .build();
-    }
-
-    public ExerciseCancelDTO.Response toCancelResponse(Exercise exercise, Guest guest) {
-        return ExerciseCancelDTO.Response.builder()
-                .memberName(guest.getGuestName())
-                .currentParticipants(exercise.getNowCapacity())
-                .build();
-    }
-
-    public ExerciseDeleteDTO.Response toDeleteResponse(Exercise exercise) {
-        return ExerciseDeleteDTO.Response.builder()
-                .deletedExerciseId(exercise.getId())
-                .build();
-    }
-
-    public ExerciseUpdateDTO.Response toUpdateResponse(Exercise exercise) {
-        return ExerciseUpdateDTO.Response.builder()
-                .exerciseId(exercise.getId())
-                .updatedAt(exercise.getUpdatedAt())
-                .build();
-    }
-
     public ExerciseDetailDTO.Response toDetailResponse(
             boolean isManager,
             ExerciseDetailDTO.ExerciseInfo exerciseInfo,
