+++ src/main/java/umc/cockple/demo/domain/exercise/converter/ExerciseLifecycleMapper.java
@@ -1,0 +1,77 @@
+package umc.cockple.demo.domain.exercise.converter;
+
+import org.springframework.stereotype.Component;
+import umc.cockple.demo.domain.exercise.domain.Exercise;
+import umc.cockple.demo.domain.exercise.dto.ExerciseCreateDTO;
+import umc.cockple.demo.domain.exercise.dto.ExerciseDeleteDTO;
+import umc.cockple.demo.domain.exercise.dto.ExerciseUpdateDTO;
+
+@Component
+public class ExerciseLifecycleMapper {
+
+    public ExerciseCreateDTO.Command toCreateCommand(ExerciseCreateDTO.Request request) {
+        return ExerciseCreateDTO.Command.builder()
+                .date(request.toParsedDate())
+                .startTime(request.toParsedStartTime())
+                .endTime(request.toParsedEndTime())
+                .maxCapacity(request.maxCapacity())
+                .partyGuestAccept(request.allowMemberGuestsInvitation())
+                .outsideGuestAccept(request.allowExternalGuests())
+                .notice(request.notice())
+                .build();
+    }
+
+    public ExerciseCreateDTO.AddrCommand toAddrCreateCommand(ExerciseCreateDTO.Request request) {
+        return ExerciseCreateDTO.AddrCommand.builder()
+                .roadAddress(request.roadAddress())
+                .buildingName(request.buildingName())
+                .latitude(request.latitude())
+                .longitude(request.longitude())
+                .build();
+    }
+
+    public ExerciseUpdateDTO.Command toUpdateCommand(ExerciseUpdateDTO.Request request) {
+        return ExerciseUpdateDTO.Command.builder()
+                .date(request.toParsedDate())
+                .startTime(request.toParsedStartTime())
+                .endTime(request.toParsedEndTime())
+                .maxCapacity(request.maxCapacity())
+                .partyGuestAccept(request.allowMemberGuestsInvitation())
+                .outsideGuestAccept(request.allowExternalGuests())
+                .notice(request.notice())
+                .build();
+    }
+
+    public ExerciseUpdateDTO.AddrCommand toAddrUpdateCommand(ExerciseUpdateDTO.Request request) {
+        if (request.roadAddress() != null || request.buildingName() != null ||
+                request.latitude() != null || request.longitude() != null) {
+            return ExerciseUpdateDTO.AddrCommand.builder()
+                    .roadAddress(request.roadAddress())
+                    .buildingName(request.buildingName())
+                    .latitude(request.latitude())
+                    .longitude(request.longitude())
+                    .build();
+        }
+        return null;
+    }
+
+    public ExerciseCreateDTO.Response toCreateResponse(Exercise exercise) {
+        return ExerciseCreateDTO.Response.builder()
+                .exerciseId(exercise.getId())
+                .createdAt(exercise.getCreatedAt())
+                .build();
+    }
+
+    public ExerciseDeleteDTO.Response toDeleteResponse(Exercise exercise) {
+        return ExerciseDeleteDTO.Response.builder()
+                .deletedExerciseId(exercise.getId())
+                .build();
+    }
+
+    public ExerciseUpdateDTO.Response toUpdateResponse(Exercise exercise) {
+        return ExerciseUpdateDTO.Response.builder()
+                .exerciseId(exercise.getId())
+                .updatedAt(exercise.getUpdatedAt())
+                .build();
+    }
+}
